extract drag input into separate file
