+++ src/input.js
@@ -1,7 +1,7 @@
-import { getIsGltf, getIsGlb } from './utils.js';
 import { gltfMouseInput } from './mouse_input.js';
 import { gltfTouchInput } from './touch_input.js';
 import { gltfKeyboardInput } from './keyboard_input.js';
+import { gltfDragInput } from './drag_input.js';
 
 class gltfInput
 {
@@ -10,6 +10,7 @@ class gltfInput
         this.mouseInput = new gltfMouseInput(canvas);
         this.touchInput = new gltfTouchInput();
         this.keyboardInput = new gltfKeyboardInput();
+        this.dragInput = new gltfDragInput();
 
         this.onZoom = () => { };
         this.onRotate = () => { };
@@ -22,6 +23,7 @@ class gltfInput
         this.mouseInput.onPan = ((x, y) => this.onPan(x, y)).bind(this);
         this.touchInput.onRotate = ((x, y) => this.onRotate(x, y)).bind(this);
         this.keyboardInput.onResetCamera = (() => this.onResetCamera()).bind(this);
+        this.dragInput.onDropFiles = ((file, additionalFiles) => this.onDropFiles(file, additionalFiles)).bind(this);
     }
 
     setupGlobalInputBindings(document)
@@ -29,6 +31,7 @@ class gltfInput
         this.mouseInput.setupGlobalInputBindings(document);
         this.touchInput.setupGlobalInputBindings(document);
         this.keyboardInput.setupGlobalInputBindings(document);
+        this.dragInput.setupGlobalInputBindings(document);
     }
 
     setupCanvasInputBindings(canvas)
@@ -36,42 +39,7 @@ class gltfInput
         this.mouseInput.setupCanvasInputBindings(canvas);
         this.touchInput.setupCanvasInputBindings(canvas);
         this.keyboardInput.setupCanvasInputBindings(canvas);
-
-        canvas.ondrop = this.dropEventHandler.bind(this);
-        canvas.ondragover = this.dragOverHandler.bind(this);
-    }
-
-    // for some reason, the drop event does not work without this
-    dragOverHandler(event)
-    {
-        event.preventDefault();
-    }
-
-    dropEventHandler(event)
-    {
-        event.preventDefault();
-
-        let additionalFiles = [];
-        let mainFile;
-        for (const file of event.dataTransfer.files)
-        {
-            if (getIsGltf(file.name) || getIsGlb(file.name))
-            {
-                mainFile = file;
-            }
-            else
-            {
-                additionalFiles.push(file);
-            }
-        }
-
-        if (mainFile === undefined)
-        {
-            console.warn("No gltf/glb file found. Provided files: " + additionalFiles.map(f => f.name).join(", "));
-            return;
-        }
-
-        this.onDropFiles(mainFile, additionalFiles);
+        this.dragInput.setupCanvasInputBindings(canvas);
     }
 }
